scripts/requirements.txt
# --- AI & Language (The Brain) ---
google-generativeai==0.8.6
pydantic==2.12.5

# --- Computer Vision (The Eyes) ---
mediapipe>=0.10.33
opencv-contrib-python==4.13.0.92
# --- Communication (The Bridge) ---
python-osc==1.8.3

# --- Utilities ---
numpy==2.0.2
matplotlib==3.9.4
pygame-ce>=2.5.0